Add PR templates for projects moving levels
These are the results of the projects-moving-level taskforce items that were compiled by emily into one document. I've gone ahead and taken those items and split it into three documents, one for sandbox, one for incubating, and one for graduation. 

The idea is that a project would file an tracking issue, and then would PR in one of these templates for their evaluation, and then the TOC would fill in the blanks for each of the criteria items. Happy to make changes as needed!

Signed-off-by: Jorge O. Castro <jorge.castro@gmail.com>